rename descriptor_validate to descriptor_validation
